修改微信智能相关的bean
