+++ src/main/java/weixin/popular/bean/semantic/semproxy/inner/EndLoc.java
@@ -1,0 +1,10 @@
+package weixin.popular.bean.semantic.semproxy.inner;
+
+/**
+ * @program: weixin-popular
+ * @description:
+ * @author: 01
+ * @create: 2018-08-18 13:34
+ **/
+public class EndLoc extends StartLoc {
+}
