+++ src/main/java/weixin/popular/bean/semantic/semproxy/inner/Semantic.java
@@ -1,0 +1,28 @@
+package weixin.popular.bean.semantic.semproxy.inner;
+
+/**
+ * @program: weixin-popular
+ * @description:
+ * @author: 01
+ * @create: 2018-08-18 13:30
+ **/
+public class Semantic {
+    private Details details;
+    private String intent;
+
+    public Details getDetails() {
+        return details;
+    }
+
+    public void setDetails(Details details) {
+        this.details = details;
+    }
+
+    public String getIntent() {
+        return intent;
+    }
+
+    public void setIntent(String intent) {
+        this.intent = intent;
+    }
+}
